+++ core-compiler/src/test/java/eu/f3rog/blade/compiler/arg/InnerClassTest.java
@@ -49,7 +49,7 @@ public final class InnerClassTest extends BaseTest {
                 .body(
                         "abstract class $T {",
                         "",
-                        "   @Weave(into=\"^onCreate\", args = {\"android.os.Bundle\"}, statement = \"com.example.$T.inject(this);\")",
+                        "   @Weave(into=\"0^onCreate\", args = {\"android.os.Bundle\"}, statement = \"com.example.$T.inject(this);\")",
                         "   public static void inject(Wrapper.MyFragment target) {",
                         "       if (target.getArguments() == null) {",
                         "           return;",